feat: add Dynamic and DisposeBag
This commit adds the Dynamic class to the project. This is a special generic type that allows any property stored in it to be observed. It also adds the DisposeBag class which allows any object to store a function observing a dynamic locally. This makes managing the memory taken up by those observers much easier by cleaning it up when the object with the dispose bag is dereferenced.